+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * 24.x.x
   - New Features:
+    - Added SVRG and LSVRG stochastic functions (#1625)
     - Added SAG and SAGA stochastic functions (#1624)
     - Allow `SumFunction` with 1 item (#1857)
   - Enhancements: